_archive-20251009-105732/CV-Reformatter-POC-SafeCleanup/requirements.txt
python-docx>=1.1.2
pytest>=8.0.0
mammoth>=1.6.0
